Handle setting references separately.
Honestly, this was an "eureka!" moment for me.
I already had an idea for re-factoring but this way to handle setting
references is so easy and so simple.

I also renamed references -> refs because my fingers got tired, and I
extracted the constants into pattern matches. This will happen with
everything else soon.